specified the min row group
The value of the parameter min_row_group was not specified in the pyarrow-function write_to_table(). The default is None, which might split the file into many small row groups, reducing the effect of the compression, resulting in lower performance. After some testing I decided to set the min value to 500 000.